feat: Enhance quote handling with ABI encoding and intent request generation
