+++ crates/solver-discovery/Cargo.toml
@@ -21,7 +21,7 @@ hex = "0.4"
 reqwest = "0.12"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-solver-types = { path = "../solver-types" }
+solver-types = { path = "../solver-types", features = ["abi"] }
 thiserror = "1.0"
 tokio = { version = "1.0", features = ["sync", "rt-multi-thread"] }
 toml = { workspace = true }
