+++ crates/solver-service/Cargo.toml
@@ -31,7 +31,7 @@ solver-order = { path = "../solver-order" }
 solver-pricing = { path = "../solver-pricing" }
 solver-settlement = { path = "../solver-settlement" }
 solver-storage = { path = "../solver-storage" }
-solver-types = { path = "../solver-types" }
+solver-types = { path = "../solver-types", features = ["abi"] }
 thiserror = "1.0"
 tokio = { version = "1.0", features = ["full"] }
 toml = { workspace = true }
